core: Trait 'ConnectionInfo'
This trait is used to retrieve local/remote multiaddr of a connection. It is bound with Transport::Output.

As a result, any I/O related with Transport has to implemente this trait, including SecioOutput, Yamux, Mplex, etc.

For convenience, ConnectionInfo depends on 'Send'.

Pnet seems to have some problem of code placement, TBD